admin/requirements.txt
wheel==0.33.6
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.7
